i've been sloppy this is after error pages for bad tokens or already voted

+++ election1/ballot/view.py
@@ -4,14 +4,32 @@ from election1.ballot.form import (CandidateForm, OfficeForm,
                                    ClassgrpForm, Candidate_reportForm, DatesForm)
 from election1.models import Classgrp, Office, Candidate, Dates
 from election1.extensions import db
-from datetime import datetime
 from sqlalchemy.exc import SQLAlchemyError
+import logging
+from flask_login import current_user
+from datetime import datetime
 
 ballot = Blueprint('ballot', __name__)
+logger = logging.getLogger(__name__)
+
+
+def classgrp_query():
+    return_list = []
+    classgrp_data = db.session.query(Classgrp).order_by(Classgrp.sortkey)
+    for c in classgrp_data:
+        return_list.append(tuple((c.id_classgrp, c.name)))
+    return return_list
 
 
 @ballot.route('/office', methods=['POST', 'GET'])
 def office():
+    if check_dates() is False:
+        flash('Please set the Election Dates before adding an office', category='danger')
+        return redirect(url_for('mains.homepage'))
+    if after_start_date():
+        flash('You cannot add, delete or edit an office after the voting start time or Election Dates are '
+              'empty', category='danger')
+        return redirect(url_for('mains.homepage'))
     office_form = OfficeForm()
     if office_form.validate_on_submit():
         office_title = request.form['office_title']
@@ -22,13 +40,15 @@ def office():
         try:
             db.session.add(new_office)
             db.session.commit()
-            # flash('successfully updates record', category='success')
+            logger.info(
+                'user ' + str(current_user.user_so_name) + ' has created the office titled ' + office_title)
             office_form.office_title.data = ''
             office_form.sortkey.data = None
             offices = Office.query.order_by(Office.sortkey)
             return render_template('office.html', form=office_form, offices=offices)
         except Exception as e:
             db.session.rollback()
+            logger.info('There is an error ' + str(e) + ' while creating the office titled ' + office_title)
             flash('There was a problem inserting record ' + str(e), category='danger')
             office_form.office_title.data = ''
             office_form.sortkey.data = None
@@ -49,13 +69,22 @@ def office():
 
 @ballot.route('/deleteoffice/<int:xid>', methods=['POST', 'GET'])
 def deleteoffice(xid):
-    office_to_delete = Office.query.get_or_404(xid)
+    if check_dates() is False:
+        flash('Please set the Election Dates before deleting an office', category='danger')
+        return redirect(url_for('mains.homepage'))
+    if after_start_date():
+        flash('You cannot delete an office after the voting start time or Election Dates are empty ', category='danger')
+        return redirect(url_for('mains.homepage'))
+    office_to_delete = Office.query.get(xid)
     form = CandidateForm()
     if request.method == 'POST':
 
         try:
             db.session.delete(office_to_delete)
             db.session.commit()
+            logger.info(
+                'user ' + str(
+                    current_user.user_so_name) + ' has deleted the office titled ' + office_to_delete.office_title)
             flash('successfully deleted record', category='success')
             return redirect('/office')
         except SQLAlchemyError as e:
@@ -65,13 +94,19 @@ def deleteoffice(xid):
     else:
 
         candidates = db.session.query(Candidate, Classgrp, Office).select_from(Candidate).join(Classgrp).join(
-            Office).order_by(Classgrp.sortkey, Office.sortkey).where(Candidate.id_office == id)
+            Office).order_by(Classgrp.sortkey, Office.sortkey).where(Candidate.id_office == xid)
         return render_template('office_candidate_delete.html', form=form, candidates=candidates,
                                office_to_delete=office_to_delete)
 
 
 @ballot.route('/updateoffice/<int:xid>', methods=['GET', 'POST'])
 def updateoffice(xid):
+    if check_dates() is False:
+        flash('Please set the Election Dates before updating an office', category='danger')
+        return redirect(url_for('mains.homepage'))
+    if after_start_date():
+        flash('You cannot edit an office after the voting start time or Election Dates are empty ', category='danger')
+        return redirect(url_for('mains.homepage'))
     office_form = OfficeForm()
     office_to_update = Office.query.get_or_404(xid)
     print(office_to_update.office_title)
@@ -80,6 +115,9 @@ def updateoffice(xid):
         office_to_update.sortkey = request.form['sortkey']
         try:
             db.session.commit()
+            logger.info(
+                'user ' + str(
+                    current_user.user_so_name) + ' has edited the office titled ' + office_to_update.office_title)
             flash('successfully updates record', category='success')
             office_form.office_title.data = ''
             office_form.sortkey.data = ''
@@ -93,13 +131,19 @@ def updateoffice(xid):
             offices = Office.query.order_by(Office.sortkey)
             return render_template('office.html', form=office_form, offices=offices)
     else:
-        print("gere")
         return render_template('update_office.html', form=office_form,
                                office_to_update=office_to_update)
 
 
 @ballot.route('/classgrp', methods=['POST', 'GET'])
 def classgrp():
+    if check_dates() is False:
+        flash('Please set the Election Dates before adding a class or group', category='danger')
+        return redirect(url_for('mains.homepage'))
+    if after_start_date():
+        flash('You cannot add or delete a class or a group after the voting start time or Election Dates are empty ',
+              category='danger')
+        return redirect(url_for('mains.homepage'))
     classgrp_form = ClassgrpForm()
     if classgrp_form.validate_on_submit():
         classgrp_name = request.form['name']
@@ -117,11 +161,20 @@ def classgrp():
     classgrp_form.name.data = ''
     classgrp_form.sortkey.data = None
     classgrps = Classgrp.query.order_by(Classgrp.sortkey)
+
     return render_template('classgrp.html', form=classgrp_form, classgrps=classgrps)
 
 
 @ballot.route('/deleteclass/<int:xid>')
 def deleteclass(xid):
+    if check_dates() is False:
+        flash('Please set the Election Dates before deleting a class or group', category='danger')
+        return redirect(url_for('mains.homepage'))
+    if after_start_date():
+        flash('You cannot delete an class or group after the voting start time or Election Dates are empty ',
+              category='danger')
+        return redirect(url_for('mains.homepage'))
+
     classgrp_to_delete = Classgrp.query.get_or_404(xid)
 
     try:
@@ -137,6 +190,13 @@ def deleteclass(xid):
 
 @ballot.route('/updateclass/<int:xid>', methods=['GET', 'POST'])
 def updateclass(xid):
+    if check_dates() is False:  # Check if the dates are set
+        flash('Please set the Election Dates before updating a class or group', category='danger')
+        return redirect(url_for('mains.homepage'))
+    if after_start_date():
+        flash('You cannot edit a class or group after the voting start time or Election Dates are empty ',
+              category='danger')
+        return redirect(url_for('mains.homepage'))
     classgrp_form = ClassgrpForm()
     classgrp_to_update = Classgrp.query.get_or_404(xid)
     print(classgrp_to_update.name)
@@ -158,7 +218,6 @@ def updateclass(xid):
             classgrps = Classgrp.query.order_by(Classgrp.sortkey)
             return render_template('classgrp.html', form=classgrp_form, classgrps=classgrps)
     else:
-        print("gere")
         return render_template('update_classgrp.html', form=classgrp_form,
                                classgrp_to_update=classgrp_to_update)
 
@@ -172,7 +231,7 @@ def candidate_report():
         choices_office = request.form['choices_office']
 
         # the first if determines if the choice_office is an int 0
-        # I built the candidate_report.html to add a choice of 'All Offices' which is not in the BD
+        # I built the candidate_report.html to add a choice of 'All Offices' which is not in the DB
         # I pass list_of_offices to the html and build the select offices manually fo this html
         if int(choices_office) == 0:
             candidates = db.session.query(Candidate, Classgrp, Office).select_from(Candidate).join(Classgrp).join(
@@ -186,17 +245,32 @@ def candidate_report():
                                    form=form, candidates=candidates, list_of_offices=list_of_offices)
 
     else:
+        form.choices_classgrp.choices = classgrp_query()
         return render_template("candidate_report.html", form=form, list_of_offices=list_of_offices)
 
 
 @ballot.route('/candidate', methods=['GET', 'POST'])
 def candidate():
+    if check_dates() is False:
+        flash('Please set the Election Dates before adding a candidate', category='danger')
+        return redirect(url_for('mains.homepage'))
+    if after_start_date():
+        flash('You cannot add or delete a candidate after the voting start time or Election Dates are empty ',
+              category='danger')
+        return redirect(url_for('mains.homepage'))
     form = CandidateForm()
     if request.method == 'POST':
         firstname = request.form['firstname']
         lastname = request.form['lastname']
         choices_classgrp = request.form['choices_classgrp']
         choices_office = request.form['choices_office']
+
+        # Check if a valid option is selected
+        if choices_classgrp == "Please select":
+            flash('Please select a valid option for class', category='danger')
+            form.choices_classgrp.choices = classgrp_query()
+            return render_template('candidate.html', form=form)
+
         new_candidate = Candidate(firstname=firstname,
                                   lastname=lastname,
                                   id_classgrp=choices_classgrp,
@@ -205,16 +279,22 @@ def candidate():
         try:
             db.session.add(new_candidate)
             db.session.commit()
-            # flash('candidate added successfully',category='success')
             return redirect(url_for('ballot.candidate'))
         except SQLAlchemyError as e:
             db.session.rollback()
             flash('problem adding candidate ' + str(e), category='danger')
             return redirect('/candidate')
     else:
-        candidates = db.session.query(Candidate, Classgrp, Office).select_from(Candidate).join(Classgrp).join(
-            Office).order_by(Classgrp.sortkey, Office.sortkey)
-        return render_template('candidate.html', form=form, candidates=candidates)
+        form.choices_classgrp.choices = classgrp_query()
+        return render_template('candidate.html', form=form)
+
+
+@ballot.route('/candidate/search')
+def candidate_search():
+    group = request.args.get('choices_classgrp', type=int)
+    candidates = db.session.query(Candidate, Classgrp, Office).select_from(Candidate).join(Classgrp).join(
+        Office).order_by(Classgrp.sortkey, Office.sortkey).where(Classgrp.id_classgrp == group)
+    return render_template('candidate_search_results.html',  candidates=candidates)
 
 
 @ballot.route('/deletecandidate/<int:xid>')
@@ -235,20 +315,8 @@ def deletecandidate(xid):
 @ballot.route('/dates', methods=['GET', 'POST'])
 def dates():
     form = DatesForm()
-    print(DatesForm.errors)
     if request.method == 'POST':
         if form.validate_on_submit():
-            # if request.method == 'POST':
-            # these worked with mySQL but not SQLite
-            # SQLite does not have a datetime column used Integer and epoch
-            #
-            # start_date_time = request.form.get('start_date_time')
-            # end_date_time = request.form.get('end_date_time')
-            # new_dates = Dates(start_date_time=start_date_time,end_date_time=end_date_time)
-            # db.session.add(new_dates)
-            # db.session.commit()
-
-            #
 
             datetime_str = request.form.get('start_date_time').replace("T", " ")
             datetime_etr = request.form.get('end_date_time').replace("T", " ")
@@ -263,8 +331,17 @@ def dates():
             db.session.add(new_dates)
             db.session.commit()
 
+            # # Log the start and end dates
+            # logger.info(f'user ' + str(current_user.user_so_name) + " has added the following dates:"
+            #             "Start date: {datetime_object_start}, End date: {datetime_object_end}")
+
+            # Log the start and end dates
+            logger.info('user ' + str(current_user.user_so_name) + " has added the following dates:")
+
+            logger.info(f'Start date: {datetime_object_start}, End date: {datetime_object_end}')
+
             return redirect(url_for('mains.homepage'))
-    print('edates')
+
     edate_dict = {}
     edates = Dates.query.all()
     if edates is not None:
@@ -275,14 +352,17 @@ def dates():
                 "end": datetime.fromtimestamp(edate.end_date_time)
             }
             edate_dict.update(new_edate_dict)
-    print(edate_dict)
+
     return render_template('dates.html', form=form, edate_dict=edate_dict)
 
 
-@ballot.route('/deletedates/<int:xid>')
-def deletedates(xid):
-    date_to_delete = Dates.query.get_or_404(xid)
-
+@ballot.route('/deletedates/')
+def deletedates():
+    # Get the first set of datee in the table because we are only storing one set of dates
+    date_to_delete = Dates.query.first()
+    start_date: datetime = datetime.fromtimestamp(date_to_delete.start_date_time)
+    logger.info(
+        'user {0} has deleted the date with start date of {1}'.format(str(current_user.user_so_name), str(start_date)))
     try:
         db.session.delete(date_to_delete)
         db.session.commit()
@@ -292,3 +372,25 @@ def deletedates(xid):
         db.session.rollback()
         flash('There was a problem deleting record ' + str(e))
         return redirect('/dates')
+
+
+def after_start_date():
+    date = Dates.query.first()
+    start_date_time = datetime.fromtimestamp(date.start_date_time)
+
+    # Get the current date time
+    current_date_time = datetime.now()
+
+    # Check if current date time is less than start date time
+    if current_date_time > start_date_time:
+        return True
+    else:
+        return False
+
+
+def check_dates():
+    date = Dates.query.first()
+    if date is None:
+        return False
+    else:
+        return True
